Invert HDRPassRecorder dependencies with IHDRDrawable interface
HDRPassRecorder no longer depends on ~20 concrete rendering system types.
Instead, it depends only on the IHDRDrawable interface. Rendering systems
register themselves (via adapters) and the recorder iterates them in order.

New files:
- IHDRDrawable.h: Interface with recordHDRDraw() and shouldDraw() + HDRDrawParams
- HDRDrawableAdapters.h: Simple adapters wrapping IRecordable/IRecordableAnimated
- SceneObjectsDrawable: Extracted scene object + tree + rock rendering
- SkinnedCharDrawable: Extracted player + NPC skinned mesh rendering
- DebugLinesDrawable: Extracted debug line rendering with viewport setup
- WaterDrawable: Water rendering with temporal tile cull conditional

HDRPassRecorder.cpp reduced from 676 to 93 lines. Registration happens in
Renderer::createHDRPassRecorder() with draw order and slot assignments.

Also fixes pre-existing powf build error in GuiGrassTab.cpp.

+++ src/gui/GuiGrassTab.cpp
@@ -2,6 +2,7 @@
 #include "core/interfaces/IGrassControl.h"
 #include "vegetation/GrassConstants.h"
 #include <imgui.h>
+#include <cmath>
 
 void GuiGrassTab::render(IGrassControl& grass) {
     // System Overview
@@ -77,7 +78,7 @@ void GuiGrassTab::render(IGrassControl& grass) {
         } else {
             float t = (dist - GrassConstants::CULL_START_DISTANCE) /
                       (GrassConstants::CULL_END_DISTANCE - GrassConstants::CULL_START_DISTANCE);
-            values[i] = 1.0f - powf(t, 1.0f / GrassConstants::CULL_POWER);
+            values[i] = 1.0f - std::pow(t, 1.0f / GrassConstants::CULL_POWER);
         }
     }
     ImGui::PlotLines("##DensityFalloff", values, numSamples, 0, nullptr, 0.0f, 1.0f, ImVec2(-1, 60));
